Rename y-redis testing files to redis for consistency and update import paths
