+++ src/infra/redis/testing/stream-messages-reply.factory.ts
@@ -1,5 +1,5 @@
 import { Factory } from 'fishery';
-import { StreamMessagesReply } from '../../../infra/redis/interfaces/index.js';
+import { StreamMessagesReply } from '../interfaces/index.js';
 import { streamMessageReplyFactory } from './stream-message-reply.factory.js';
 
 export const streamMessagesReplyFactory = Factory.define<StreamMessagesReply[]>(({ sequence }) => {
